Improve handing of auth and user names when updating.
Usernames can contain alphanumeric characters, _, - and . on the gui. The forgejo cli actually fails with a dot in the username despite theoretically allowing it.
Auth names can have any character in them.

This change improves the regular expressions used to decide whether to create or update users and auth methods to deal with more of these character sets.

Closes #18