Generate jinja templates from app.example.ini
A script, update_from_default_ini.py reads the `app.example.ini` file from
a git clone of forgejo's repository, uses heuristics based on comment structure
and some hardcoded exceptions to to define whether an option is a boolean,
string or list and generates the magority of jinja templates.

Some sections cannot be dealt with this automated way and are dealt with
custom templates in forgejo.custom.d.
The script introduces the notion that some options are deprecated and this is
currently only used in template comments.